+++ R/utilities-pk-analysis.R
@@ -7,10 +7,10 @@
 #'
 #' @examples
 #'
-#' simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
+#' simPath <- system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
 #' sim <- loadSimulation(simPath)
 #'
-#' addOutputs("Organism|VenousBlood|*|Caffeine", sim)
+#' addOutputs("Organism|VenousBlood|*|Aciclovir", sim)
 #' results <- runSimulations(sim)[[1]]
 #' pkAnalyses <- calculatePKAnalyses(results)
 #' @export